+++ src/cpu/ref_eltwise.cpp
@@ -190,7 +190,7 @@ void ref_eltwise_fwd_t<data_type>::execute_forward_dense() {
     dst += data_d.blocking_desc().offset_padding;
 
 #   pragma omp parallel for schedule(static)
-    for (int e = 0; e < nelems; ++e) {
+    for (size_t e = 0; e < nelems; ++e) {
         const data_t s = src[e];
         data_t &d = dst[e];
 
@@ -277,7 +277,7 @@ void ref_eltwise_bwd_t<data_type>::execute_backward_dense() {
     diff_src += diff_data_d.blocking_desc().offset_padding;
 
 #   pragma omp parallel for schedule(static)
-    for (int e = 0; e < nelems; ++e) {
+    for (size_t e = 0; e < nelems; ++e) {
         const data_t dd = diff_dst[e];
         const data_t s = src[e];
         data_t &ds = diff_src[e];
